Fix WKT parsing bug in doesCross180() with GDAL 3.
Notes by Jim Klassen:

In GDAL 3.x `geom.ExportToWkt()` may return coordinates formatted as
integers (in the case that they actually fall on an integer).  Previous
GDAL versions appear to have always formatted the coordinates as floats
(with a decimal point).  This broke the regex used to parse the
x-coordinates out of the WKT.

For example, the `-72` below was being skipped by the regex:
```
POLYGON ((
-102.439166666667 -72.3713888888889 0,
-103.113055555556 -72.3983333333333 0,
-103.104166666667 -72.0288888888889 0,
-102.446388888889 -72 0,
-102.439166666667 -72.3713888888889 0
))
```

Instead of exporting to WKT and trying to parse the WKT into floats,
loop over the rings of the polygon and then the points of the rings
to extract the floating point X-coordinates directly using GetX().